+++ src/string.ts
@@ -102,6 +102,34 @@ function isBase64(s: any, urlEncoded = false): boolean {
   return isString(s, true) && regex.test(s);
 }
 
+
+// b64Reg regex explained :
+// ^ : The caret symbol (^) at the beginning of the pattern asserts the position at the start of the string. 
+// This means the pattern will only match if the string starts with the specified characters.
+
+// [A-Za-z0-9\-_] : The square brackets define a character set that includes uppercase letters (A-Z), 
+// lowercase letters (a-z), digits (0-9), hyphens (-), and underscores (_). 
+// The pattern will match any single character that is within this set.
+
+// + : The plus sign following the character set indicates that the preceding element must appear one or more times.
+// This means the string must contain at least one of the specified characters and can contain many.
+
+// = : The equals sign is a literal character that must appear in the string.
+// It is not part of a character set or a special regex symbol, 
+// so it matches the equals sign character exactly.
+
+// {0,2}: The curly braces specify that the preceding equals sign can appear zero, one, or two times.
+// This means the string can end with up to two equals signs, but it is not required to have any.
+
+// $: The dollar sign at the end of the pattern asserts the position at the end of the string.
+// This ensures that the entire string must match the pattern from start to finish, 
+// with no extra characters before or after.
+
+// In summary, this regex pattern matches strings that start with one or more alphanumeric characters, 
+// hyphens, or underscores, and optionally end with up to two equals signs. 
+// This type of pattern is often used for validating encoded strings, such as Base64 URL-safe encoded strings, 
+// which may include padding characters (=) at the end.
+
 const b64Reg = /^[A-Za-z0-9\-_]+={0,2}$/;
 function isJWT(t: any): t is string {
   if (!isString(t, true))